\name{print.mtnardl}
\alias{print.mtnardl}
\title{Print Method for mtnardl Objects}
\usage{
\method{print}{mtnardl}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"mtnardl"}.}

\item{digits}{Integer. Number of digits to display. Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for mtnardl Objects
}
