\name{summary.mtnardl}
\alias{summary.mtnardl}
\title{Summary Method for mtnardl Objects}
\usage{
\method{summary}{mtnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mtnardl"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary Method for mtnardl Objects
}
